requirements.txt
streamlit
Pillow
numpy
opencv-python
mediapipe
torch
torchvision
transformers
diffusers
accelerate
safetensors
